Enabled some of GF Embedded Maven plugin tests Until websocket tests, which I only tried to refactor. Signed-off-by:Ondro Mihalyi <mihalyi@omnifish.ee>
