+++ appserver/tests/embedded/maven-plugin/applicationDispatcher/src/test/java/org/glassfish/tests/applicationDispatcher/WebTest.java
@@ -21,12 +21,8 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 
 import java.io.BufferedReader;
-import java.io.File;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.net.Socket;
 import java.net.URL;
 import java.net.URLConnection;
