--- appserver/tests/embedded/maven-plugin/scattered/src/main/java/test/ServletTest3.java
@@ -1,75 +1,0 @@
-/*
- * Copyright (c) 1997, 2018 Oracle and/or its affiliates. All rights reserved.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v. 2.0, which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- *
- * This Source Code may also be made available under the following Secondary
- * Licenses when the conditions for such availability set forth in the
- * Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
- * version 2 with the GNU Classpath Exception, which is available at
- * https://www.gnu.org/software/classpath/license.html.
- *
- * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
- */
-
-package test;
-
-import java.io.*;
-import java.net.*;
-import java.util.*;
-import jakarta.servlet.*;
-import jakarta.servlet.http.*;
-import org.xml.sax.InputSource;
-import org.w3c.dom.Element;
-import org.w3c.dom.Document;
-import org.w3c.dom.NodeList;
-import org.w3c.dom.Node;
-
-import org.xml.sax.SAXException;
-import org.xml.sax.SAXParseException;
-import org.xml.sax.SAXException;
-
-// jaxp 1.0.1 imports
-import javax.xml.parsers.DocumentBuilderFactory;
-import javax.xml.parsers.DocumentBuilder;
-
-import java.net.URL;
-import java.util.ArrayList;
-import java.util.HashMap;
-
-import javax.naming.*;
-import javax.sql.*;
-import java.sql.*;
-
-public class ServletTest3 extends HttpServlet{
-
-    private ServletContext context;
-
-    public void init(ServletConfig config) throws ServletException {
-        super.init(config);
-        System.out.println("[Servlet3.init]");
-        context = config.getServletContext();
-    }
-
-    public void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        System.out.println("[Servlet3.doGet]");
-        doPost(request, response);
-    }
-
-    public void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        System.out.println("[Servlet3.doPost]");
-
-        response.setContentType("text/html");
-        PrintWriter out = response.getWriter();
-
-        out.println("FILTER-REQUEST:" + request.getSession().getAttribute("FILTER-REQUEST"));
-        out.println("FILTER-FORWARD:" + request.getSession().getAttribute("FILTER-FORWARD"));
-        out.println("FILTER-INCLUDE:" + request.getSession().getAttribute("FILTER"));
-    }
-
-}
-
-
-
