+++ appserver/tests/embedded/maven-plugin/websockets/src/main/webapp/index.html
@@ -57,5 +57,8 @@
             </div>
         </div>
         <iframe id="websockets-frame" style="display: none;"></iframe>
+        <script type="text/javascript">
+            app.initialize();
+        </script>
     </body>
 </html>
