+++ appserver/tests/embedded/pom.xml
@@ -40,8 +40,8 @@
         <module>glassfish_resources_xml</module>
         <module>ejb</module>
         <module>scatteredarchive</module>
-        <!--<module>maven-plugin</module>
-        <module>web</module> -->
+        <module>maven-plugin</module>
+        <!--<module>web</module> -->
     </modules>
 
     <dependencyManagement>
